Relocate MAX_DISTINCT_VALUES_FOR_AUTO_ADD to shared constants module
Moves the constant out of target_respondent_helpers (which imports
from respondent_service) into a neutral service_layer/constants.py.
Side effect: respondent_field_schema_service can now import it at
the top level without the circular dance that was required before.

Target-related modules (assembly_service, targets blueprint,
targets_legacy blueprint, and the integration test) now import
from the new location directly. target_respondent_helpers re-exports
the constant through its __all__ so the symbol stays available for
any caller still reaching for it there.
