+++ backend/tests/integration/test_targets_respondent_data.py
@@ -11,9 +11,9 @@ from opendlp.domain.assembly import Assembly
 from opendlp.domain.respondents import Respondent
 from opendlp.domain.users import User
 from opendlp.domain.value_objects import GlobalRole
-from opendlp.entrypoints.blueprints.targets_legacy import MAX_DISTINCT_VALUES_FOR_AUTO_ADD
 from opendlp.entrypoints.flask_app import create_app
 from opendlp.service_layer import assembly_service
+from opendlp.service_layer.constants import MAX_DISTINCT_VALUES_FOR_AUTO_ADD
 from opendlp.service_layer.respondent_service import (
     get_respondent_attribute_value_counts,
 )
